finish debug

+++ python/tests/op_mappers/test_argmax_op.py
@@ -104,7 +104,7 @@ class TestArgmaxCase3(TestArgmaxOp):
     def init_input_data(self):
         self.axis = 1
         self.shape = [2, 3, 4]
-        self.input_dtype = "float64"
+        self.input_dtype = "float32"
         self.output_dtype = "int64"
         self.flatten = False
         self.keepdims = False
